// src/model/currency.ts
// https://bankenverband.de/service/waehrungsrechner-bdb-link/wahrungen-und-abkurzungen/
export enum Currency {
    USD = 'USD',
    EUR = 'EUR',
    JPY = 'JPY',
    MXN = 'MXN',
    GBP = 'GBP',
    CAD = 'CAD',
    AUD = 'AUD',
    CHF = 'CHF',
    CNY = 'CNY',
    INR = 'INR',
    SEK = 'SEK',
    NZD = 'NZD',
    BRL = 'BRL',
    SGD = 'SGD',
    HKD = 'HKD',
    NOK = 'NOK',
    ZAR = 'ZAR',
    RUB = 'RUB',
    TRY = 'TRY',
    KRW = 'KRW',
    IDR = 'IDR',
    MYR = 'MYR',
    PHP = 'PHP',
    THB = 'THB',
    PLN = 'PLN',
    CZK = 'CZK',
    ILS = 'ILS',
    CLP = 'CLP',
    AED = 'AED',
    SAR = 'SAR',
    TWD = 'TWD'
}